# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"description":"Login user","operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"summary":"","tags":["auth"]}},"/auth/register":{"post":{"description":"Register user","operationId":"register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"","tags":["auth"]}},"/auth/refresh":{"post":{"description":"Refresh access token","operationId":"refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshResponseDto"}}}}},"summary":"","tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Logout and revoke refresh token","tags":["auth"]}},"/auth/logout-all":{"post":{"operationId":"AuthController_logoutAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Logout from all devices","tags":["auth"]}},"/auth/change-password":{"post":{"description":"Change user password and re-encrypt all data","operationId":"changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"","tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get current user","tags":["auth"]}},"/auth/2fa/verify-login":{"post":{"operationId":"verifyTwoFactorLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginVerifyTotpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"summary":"Complete 2FA login","tags":["auth"]}},"/auth/2fa/setup":{"post":{"operationId":"setupTwoFactor","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFactorSetupResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Generate 2FA setup (QR code + secret)","tags":["auth"]}},"/auth/2fa/confirm":{"post":{"operationId":"confirmTwoFactor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyTotpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFactorBackupCodesResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Confirm 2FA setup with first TOTP code","tags":["auth"]}},"/auth/2fa/disable":{"post":{"operationId":"disableTwoFactor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DisableTwoFactorDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Disable 2FA (requires password)","tags":["auth"]}},"/auth/2fa/status":{"get":{"operationId":"getTwoFactorStatus","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFactorStatusResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Check if 2FA is enabled","tags":["auth"]}},"/auth/2fa/backup-codes/regenerate":{"post":{"operationId":"regenerateBackupCodes","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyTotpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFactorBackupCodesResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Regenerate backup codes (requires TOTP code)","tags":["auth"]}},"/projects":{"post":{"description":"Create a new project","operationId":"createProject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Projects"]},"get":{"description":"Get all projects","operationId":"getAllProjects","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":10,"type":"number"}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string","enum":["asc","desc"]}},{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["ACTIVE","ARCHIVED"]}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Projects"]}},"/projects/{id}":{"put":{"description":"Update a project","operationId":"updateProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Projects"]},"get":{"description":"Get a project by id","operationId":"getProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Projects"]},"delete":{"description":"Delete a project by id","operationId":"deleteProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Projects"]}},"/access":{"post":{"description":"Create access","operationId":"createAccess","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccessDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Access"]},"put":{"description":"Update access by id","operationId":"updateAccessById","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccessDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Access"]}},"/access/projects/{projectId}":{"get":{"description":"Get access by project id","operationId":"getAccessByProjectId","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":10,"type":"number"}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string","enum":["asc","desc"]}},{"name":"environment","required":false,"in":"query","schema":{"type":"string","enum":["PRODUCTION","STAGING","DEVELOPMENT"]}},{"name":"accessLevel","required":false,"in":"query","schema":{"type":"string","enum":["ADMIN","EDITOR","VIEWER"]}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Access"]}},"/access/{id}":{"get":{"description":"Get access by id","operationId":"getAccessById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Access"]},"delete":{"description":"Delete access by id","operationId":"deleteAccessById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["Access"]}},"/activity-log/access/{accessId}":{"get":{"description":"Get all activities for a specific access","operationId":"getActivitiesByAccessId","parameters":[{"name":"accessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["activity-log"]}},"/activity-log/project/{projectId}":{"get":{"description":"Get all activities for a specific project","operationId":"getActivitiesByProjectId","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["activity-log"]}},"/two-factor":{"post":{"operationId":"TwoFactorController_createTwoFactor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTwoFactorDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["two-factor"]}},"/two-factor/access/{accessId}":{"get":{"operationId":"TwoFactorController_getTwoFactorByAccessId","parameters":[{"name":"accessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["two-factor"]}},"/two-factor/{id}":{"get":{"operationId":"TwoFactorController_getTwoFactorById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["two-factor"]},"put":{"operationId":"TwoFactorController_updateTwoFactor","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTwoFactorDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["two-factor"]},"delete":{"operationId":"TwoFactorController_deleteTwoFactorById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["two-factor"]}},"/recovery-code":{"post":{"description":"Create a recovery code","operationId":"createRecoveryCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecoveryCodeDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["recovery-code"]}},"/recovery-code/two-factor/{twoFactorId}":{"get":{"description":"Get recovery codes by two factor id","operationId":"getRecoveryCodeByTwoFactorId","parameters":[{"name":"twoFactorId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["recovery-code"]}},"/recovery-code/{id}":{"get":{"description":"Get a recovery code by id","operationId":"getRecoveryCodeById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["recovery-code"]},"delete":{"description":"Delete a recovery code","operationId":"deleteRecoveryCode","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["recovery-code"]}},"/recovery-code/{id}/use":{"patch":{"description":"Use a recovery code","operationId":"useRecoveryCode","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"","tags":["recovery-code"]}},"/dashboard":{"get":{"description":"Get dashboard data","operationId":"getDashboardData","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"","tags":["Dashboard"]}}},"info":{"title":"AccessLog API","description":"","version":"1.0","contact":{}},"tags":[{"name":"api","description":""}],"servers":[{"url":"/"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"}},"required":["accessToken","refreshToken","tokenType"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"plan":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","email","plan","createdAt","updatedAt"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}}},"RefreshResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"MessageResponseDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"ChangePasswordDto":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["currentPassword","newPassword"]},"MeResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"twoFactorEnabled":{"type":"boolean"}},"required":["id","email","twoFactorEnabled"]},"LoginVerifyTotpDto":{"type":"object","properties":{"sessionToken":{"type":"string"},"code":{"type":"string"}},"required":["sessionToken","code"]},"TwoFactorSetupResponseDto":{"type":"object","properties":{"qrCodeDataUrl":{"type":"string"},"secret":{"type":"string"},"otpauthUrl":{"type":"string"}},"required":["qrCodeDataUrl","secret","otpauthUrl"]},"VerifyTotpDto":{"type":"object","properties":{"code":{"type":"string"}},"required":["code"]},"TwoFactorBackupCodesResponseDto":{"type":"object","properties":{"backupCodes":{"type":"array","items":{"type":"string"}}},"required":["backupCodes"]},"DisableTwoFactorDto":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]},"TwoFactorStatusResponseDto":{"type":"object","properties":{"enabled":{"type":"boolean"}},"required":["enabled"]},"ProjectStatus":{"type":"string","enum":["ACTIVE","ARCHIVED"]},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"clientName":{"type":"string"},"status":{"allOf":[{"$ref":"#/components/schemas/ProjectStatus"}]},"description":{"type":"string"}},"required":["name","status"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"clientName":{"type":"string"},"status":{"allOf":[{"$ref":"#/components/schemas/ProjectStatus"}]},"description":{"type":"string"}}},"AccessEnvironment":{"type":"string","enum":["PRODUCTION","STAGING","DEVELOPMENT"]},"AccessLevel":{"type":"string","enum":["ADMIN","EDITOR","VIEWER"]},"CreateAccessDto":{"type":"object","properties":{"projectId":{"type":"string"},"serviceName":{"type":"string"},"serviceUrl":{"type":"string"},"environment":{"allOf":[{"$ref":"#/components/schemas/AccessEnvironment"}]},"accessLevel":{"allOf":[{"$ref":"#/components/schemas/AccessLevel"}]},"login":{"type":"string"},"password":{"type":"string"},"notes":{"type":"string"},"owner":{"type":"string"}},"required":["projectId","serviceName","serviceUrl","environment","accessLevel","login","password","notes","owner"]},"UpdateAccessDto":{"type":"object","properties":{"id":{"type":"string"},"serviceName":{"type":"string"},"serviceUrl":{"type":"string"},"environment":{"allOf":[{"$ref":"#/components/schemas/AccessEnvironment"}]},"accessLevel":{"allOf":[{"$ref":"#/components/schemas/AccessLevel"}]},"login":{"type":"string"},"password":{"type":"string"},"notes":{"type":"string"},"owner":{"type":"string"}},"required":["id","serviceName","serviceUrl","environment","accessLevel","login","password","notes","owner"]},"CreateTwoFactorDto":{"type":"object","properties":{"accessId":{"type":"string"},"type":{"type":"string"},"enabled":{"type":"boolean"}},"required":["accessId","type","enabled"]},"UpdateTwoFactorDto":{"type":"object","properties":{"type":{"type":"string"},"enabled":{"type":"boolean"}},"required":["type","enabled"]},"CreateRecoveryCodeDto":{"type":"object","properties":{"twoFactorId":{"type":"string"},"code":{"type":"string"}},"required":["twoFactorId","code"]},"EnvironmentsDto":{"type":"object","properties":{"production":{"type":"number"},"staging":{"type":"number"},"development":{"type":"number"}},"required":["production","staging","development"]},"DashboardStatsDto":{"type":"object","properties":{"totalProjects":{"type":"number"},"totalCredentials":{"type":"number"},"environments":{"$ref":"#/components/schemas/EnvironmentsDto"},"activities":{"type":"number"}},"required":["totalProjects","totalCredentials","environments","activities"]},"DashboardProjectDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"clientName":{"type":"string"},"status":{"type":"string"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"credentialsCount":{"type":"number"}},"required":["id","name","status","createdAt","updatedAt","credentialsCount"]},"DashboardActivityDto":{"type":"object","properties":{"id":{"type":"string"},"accessId":{"type":"string"},"userId":{"type":"string"},"action":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"serviceName":{"type":"string"},"projectName":{"type":"string"}},"required":["id","accessId","userId","action","createdAt","updatedAt"]},"DashboardResponseDto":{"type":"object","properties":{"stats":{"$ref":"#/components/schemas/DashboardStatsDto"},"recentProjects":{"type":"array","items":{"$ref":"#/components/schemas/DashboardProjectDto"}},"recentActivities":{"type":"array","items":{"$ref":"#/components/schemas/DashboardActivityDto"}}},"required":["stats","recentProjects","recentActivities"]}}}}