# app/slices/setup/api/data/repositories/api/models/AccessEnvironment.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum AccessEnvironment {
    PRODUCTION = 'PRODUCTION',
    STAGING = 'STAGING',
    DEVELOPMENT = 'DEVELOPMENT',
}