# app/slices/setup/api/data/repositories/api/models/CreateProjectDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProjectStatus } from './ProjectStatus';
export type CreateProjectDto = {
    name: string;
    clientName?: string;
    status: ProjectStatus;
    description?: string;
};