# app/slices/setup/api/data/repositories/api/models/DashboardActivityDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DashboardActivityDto = {
    id: string;
    accessId: string;
    userId: string;
    action: string;
    createdAt: string;
    updatedAt: string;
    serviceName?: string;
    projectName?: string;
};