# app/slices/setup/api/data/repositories/api/models/DashboardProjectDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DashboardProjectDto = {
    id: string;
    name: string;
    clientName?: string;
    status: string;
    description?: string;
    createdAt: string;
    updatedAt: string;
    credentialsCount: number;
};