# app/slices/setup/api/data/repositories/api/models/DashboardStatsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EnvironmentsDto } from './EnvironmentsDto';
export type DashboardStatsDto = {
    totalProjects: number;
    totalCredentials: number;
    environments: EnvironmentsDto;
    activities: number;
};