# app/slices/setup/api/data/repositories/api/models/EnvironmentsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EnvironmentsDto = {
    production: number;
    staging: number;
    development: number;
};