# app/slices/setup/api/data/repositories/api/models/LoginVerifyTotpDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LoginVerifyTotpDto = {
    sessionToken: string;
    code: string;
};