# app/slices/setup/api/data/repositories/api/models/MeResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MeResponseDto = {
    id: string;
    email: string;
    twoFactorEnabled: boolean;
};