# app/slices/setup/api/data/repositories/api/models/TwoFactorBackupCodesResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TwoFactorBackupCodesResponseDto = {
    backupCodes: Array<string>;
};