# app/slices/setup/api/data/repositories/api/models/TwoFactorSetupResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TwoFactorSetupResponseDto = {
    qrCodeDataUrl: string;
    secret: string;
    otpauthUrl: string;
};