# app/slices/setup/api/data/repositories/api/models/UpdateAccessDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AccessEnvironment } from './AccessEnvironment';
import type { AccessLevel } from './AccessLevel';
export type UpdateAccessDto = {
    id: string;
    serviceName: string;
    serviceUrl: string;
    environment: AccessEnvironment;
    accessLevel: AccessLevel;
    login: string;
    password: string;
    notes: string;
    owner: string;
};