# app/slices/setup/api/data/repositories/api/models/UpdateProjectDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProjectStatus } from './ProjectStatus';
export type UpdateProjectDto = {
    name?: string;
    clientName?: string;
    status?: ProjectStatus;
    description?: string;
};